+++ example/lib/screens/quill/quill_screen.dart
@@ -7,6 +7,7 @@ import 'package:flutter_quill_extensions/flutter_quill_extensions.dart'
 import 'package:share_plus/share_plus.dart' show Share;
 
 import '../../extensions/scaffold_messenger.dart';
+import '../../spell_checker/spell_checker.dart';
 import '../shared/widgets/home_screen_button.dart';
 import 'my_quill_editor.dart';
 import 'my_quill_toolbar.dart';
@@ -38,6 +39,7 @@ class _QuillScreenState extends State<QuillScreen> {
   final _editorFocusNode = FocusNode();
   final _editorScrollController = ScrollController();
   var _isReadOnly = false;
+  var _isSpellcheckerActive = false;
 
   @override
   void initState() {
@@ -56,10 +58,28 @@ class _QuillScreenState extends State<QuillScreen> {
   @override
   Widget build(BuildContext context) {
     _controller.readOnly = _isReadOnly;
+    if (!_isSpellcheckerActive) {
+      _isSpellcheckerActive = true;
+      SpellChecker.useSpellCheckerService(
+          Localizations.localeOf(context).languageCode);
+    }
     return Scaffold(
       appBar: AppBar(
         title: const Text('Flutter Quill'),
         actions: [
+          IconButton(
+            tooltip: 'Spell-checker',
+            onPressed: () {
+              SpellCheckerServiceProvider.toggleState();
+              setState(() {});
+            },
+            icon: Icon(
+              Icons.document_scanner,
+              color: SpellCheckerServiceProvider.isServiceActive()
+                  ? Colors.red.withOpacity(0.5)
+                  : null,
+            ),
+          ),
           IconButton(
             tooltip: 'Share',
             onPressed: () {
@@ -118,7 +138,7 @@ class _QuillScreenState extends State<QuillScreen> {
         ],
       ),
       floatingActionButton: FloatingActionButton(
-        child: Icon(_isReadOnly ? Icons.lock : Icons.edit),
+        child: Icon(!_isReadOnly ? Icons.lock : Icons.edit),
         onPressed: () => setState(() => _isReadOnly = !_isReadOnly),
       ),
     );